+++ test/it/shellutils/tty_test.sh
@@ -1,0 +1,3 @@
+TestTtyNotATty() {
+    echo "" | tty
+}