+++ test/it/shellutils/yes_test.sh
@@ -1,0 +1,6 @@
+TestYesHead() {
+    yes | head -n 3
+}
+TestYesString() {
+    yes mimix | head -n 2
+}